requirements.txt
django-widget-tweaks>=1.4